simplify prompt actions component
