+++ src/components/prompt-actions.tsx
@@ -1,18 +1,9 @@
-import {
-	GlobeIcon,
-	PaperclipIcon,
-	PaperPlaneRightIcon,
-	SpinnerIcon,
-	StopIcon,
-} from "@phosphor-icons/react";
-import { useHotkey } from "@tanstack/react-hotkeys";
+import { PaperPlaneRightIcon, StopIcon } from "@phosphor-icons/react";
 import type { ChatStatus } from "ai";
-import { cn } from "~/lib/utils";
-import { modelStoreActions, useModelStore } from "~/stores/model-store";
-import { usePersistedApiKeysStore } from "~/stores/persisted-api-keys-store";
+import InputUploadButton from "./input-upload-button";
 import ModelSelector from "./model-selector";
 import { Button } from "./ui/button";
-import { Tooltip, TooltipContent, TooltipTrigger } from "./ui/tooltip";
+import WebSearchButton from "./web-search-button";
 
 type Props = {
 	attachmentCount?: number;
@@ -31,72 +22,19 @@ export default function PromptActions({
 	status,
 	stop,
 }: Props) {
-	const isWebSearchEnabled = useModelStore((store) => store.isWebSearchEnabled);
-	const selectedModel = useModelStore((store) => store.selectedModel);
-	const persistedUseOpenRouter = usePersistedApiKeysStore(
-		(store) => store.persistedUseOpenRouter,
-	);
-
-	useHotkey("Mod+Shift+S", modelStoreActions.toggleIsWebSearch);
-
 	return (
 		<div className="mt-3 flex flex-row items-center justify-between gap-2 lg:mt-4">
 			<div className="flex flex-1 flex-wrap items-center gap-2">
 				<ModelSelector />
 
-				{(selectedModel.openRouterModelId.startsWith("google") ||
-					persistedUseOpenRouter) && (
-					<Tooltip>
-						<TooltipTrigger
-							render={
-								<Button
-									className={cn(
-										"transition-all duration-300 ease-out",
-										isWebSearchEnabled ? "border-primary" : "border-border",
-									)}
-									onClick={modelStoreActions.toggleIsWebSearch}
-									type="button"
-									variant={isWebSearchEnabled ? "default" : "outline"}
-								>
-									<GlobeIcon />
-									<span className="hidden sm:inline">Search</span>
-									<span className="sm:hidden">Web</span>
-								</Button>
-							}
-						/>
-						<TooltipContent>Ctrl+Shift+S</TooltipContent>
-					</Tooltip>
-				)}
+				<WebSearchButton />
 
-				<Button
-					className={cn(
-						"transition-all duration-300 ease-out",
-						attachmentCount > 0 &&
-							"border-primary/60 bg-primary/8 text-foreground",
-					)}
+				<InputUploadButton
+					attachmentCount={attachmentCount}
 					disabled={disabled}
-					onClick={onAttachClick}
-					type="button"
-					variant="outline"
-				>
-					{isUploading ? (
-						<>
-							<SpinnerIcon className="size-4 animate-spin" />
-							<span>Uploading...</span>
-						</>
-					) : (
-						<>
-							<PaperclipIcon />
-							<span className="hidden sm:inline">Upload</span>
-							<span className="sm:hidden">File</span>
-						</>
-					)}
-					{!isUploading && attachmentCount > 0 && (
-						<span className="rounded-full bg-foreground px-1.5 py-0.5 text-[10px] leading-none text-background">
-							{attachmentCount}
-						</span>
-					)}
-				</Button>
+					isUploading={isUploading}
+					onAttachClick={onAttachClick}
+				/>
 			</div>
 
 			{status === "streaming" || status === "submitted" ? (
